Git-2324 added test for validation error and fixed imports
